The Big Commit(ment)!
Configurations Menu

Now, after clicking in the configurations menu in the payment.html page
it takes you to a menu instead of directly to the statement page.

payment.html was modified in accordance.

Statement Listed

The results of the last 10 transactions in the statements are now
presented as an ordered list.

Currency Fix

The currencies in the statement and receipt pages were formatted to
BsF., this wasn't done using the jQuery Currency libraries (as in the
payment page) since those are intended to format only inout fields.

The receipt.html, statement.html, statement.js, index.html files were
affected.
index.html was modified to include the accounting.min.js file and to ser
the basic configurations of the accounting.js libraries.