WHOA! It IS even bigger!!!
Now, this is a big commit.

"default-content-divs" to cover for the restrictions of fixing the width
of all the fields of the ui-input-text class.  Also the data-roles of
the divs containing the input text fields changed to "container" from
the previous "fieldcontain". Ttiles and buttons were also included
inside "default-content-divs" containers, so all fields, buttons and
titles changed sizes and relative locations, even though the change was
not significant.

Also relative positions expressed in pixels switched to percentages, so
the display is being optimized for multiple screen sizes, this process
is not complete.

index.html, payment.html, styles.css, and the receipt files were
affected in this commit.

Tax was set to 0.

The receipt interface now has email validation (though not email sending
yet) and the cancel button is working.

The receipt files were affected by this change.